+++ packages/builders/src/vercel-build-output-api.ts
@@ -13,12 +13,11 @@ export class VercelBuildOutputAPIBuilder extends BaseBuilder {
     await mkdir(workflowGeneratedDir, { recursive: true });
 
     const inputFiles = await this.getInputFiles();
-    const tsConfig = await this.getTsConfigOptions();
+    const tsconfigPath = await this.findTsConfigPath();
     const options = {
       inputFiles,
       workflowGeneratedDir,
-      tsBaseUrl: tsConfig.baseUrl,
-      tsPaths: tsConfig.paths,
+      tsconfigPath,
     };
     const manifest = await this.buildStepsFunction(options);
     await this.buildWorkflowsFunction(options);
@@ -39,13 +38,11 @@ export class VercelBuildOutputAPIBuilder extends BaseBuilder {
   private async buildStepsFunction({
     inputFiles,
     workflowGeneratedDir,
-    tsPaths,
-    tsBaseUrl,
+    tsconfigPath,
   }: {
     inputFiles: string[];
     workflowGeneratedDir: string;
-    tsBaseUrl?: string;
-    tsPaths?: Record<string, string[]>;
+    tsconfigPath?: string;
   }) {
     console.log('Creating Vercel Build Output API steps function');
     const stepsFuncDir = join(workflowGeneratedDir, 'step.func');
@@ -55,8 +52,7 @@ export class VercelBuildOutputAPIBuilder extends BaseBuilder {
     const { manifest } = await this.createStepsBundle({
       inputFiles,
       outfile: join(stepsFuncDir, 'index.js'),
-      tsBaseUrl,
-      tsPaths,
+      tsconfigPath,
     });
 
     // Create package.json and .vc-config.json for steps function
@@ -72,13 +68,11 @@ export class VercelBuildOutputAPIBuilder extends BaseBuilder {
   private async buildWorkflowsFunction({
     inputFiles,
     workflowGeneratedDir,
-    tsPaths,
-    tsBaseUrl,
+    tsconfigPath,
   }: {
     inputFiles: string[];
     workflowGeneratedDir: string;
-    tsBaseUrl?: string;
-    tsPaths?: Record<string, string[]>;
+    tsconfigPath?: string;
   }): Promise<void> {
     console.log('Creating Vercel Build Output API workflows function');
     const workflowsFuncDir = join(workflowGeneratedDir, 'flow.func');
@@ -87,8 +81,7 @@ export class VercelBuildOutputAPIBuilder extends BaseBuilder {
     await this.createWorkflowsBundle({
       outfile: join(workflowsFuncDir, 'index.js'),
       inputFiles,
-      tsBaseUrl,
-      tsPaths,
+      tsconfigPath,
     });
 
     // Create package.json and .vc-config.json for workflows function
@@ -102,10 +95,7 @@ export class VercelBuildOutputAPIBuilder extends BaseBuilder {
     workflowGeneratedDir,
     bundle = true,
   }: {
-    inputFiles: string[];
     workflowGeneratedDir: string;
-    tsBaseUrl?: string;
-    tsPaths?: Record<string, string[]>;
     bundle?: boolean;
   }): Promise<void> {
     console.log('Creating Vercel Build Output API webhook function');