+++ packages/utils/src/index.ts
@@ -1,2 +1,3 @@
+export { pluralize } from './pluralize.js';
 export { once, type PromiseWithResolvers, withResolvers } from './promise.js';
 export { parseDurationToDate } from './time.js';